requirements.txt
python3-tflite-runtime
libatlas-base-dev
numpy==1.26.4
librosa==0.10.1
pyserial==3.5
